Move startup code into programs.cs and fix warnings
