+++ Hippo.Core/Hippo.Core.csproj
@@ -3,13 +3,15 @@
     <TargetFramework>net8.0</TargetFramework>
     <AddRazorSupportForMvc>true</AddRazorSupportForMvc>
     <ImplicitUsings>enable</ImplicitUsings>
+    <!-- Specifying RIDs is necessary for making use of sqlite libraries -->
+    <RuntimeIdentifiers>win-x64;linux-x64;osx-x64;</RuntimeIdentifiers>
   </PropertyGroup>
   <ItemGroup>
     <FrameworkReference Include="Microsoft.AspNetCore.App" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="AggieEnterpriseApi" Version="0.2.219" />
-    <PackageReference Include="EFCore.BulkExtensions" Version="6.7.15" />
+    <PackageReference Include="EFCore.BulkExtensions" Version="8.1.3" />
     <PackageReference Include="ietws" Version="0.2.12" />
     <PackageReference Include="Microsoft.Azure.KeyVault" Version="3.0.5" />
     <PackageReference Include="Microsoft.EntityFrameworkCore" Version="8.0.14" />
